requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
oauthlib==3.2.2
openai==1.20.0
psycopg2==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.29
tqdm==4.66.2
tweepy==4.14.0
typing_extensions==4.11.0
urllib3==2.2.1
